Userproofing
But still no warranty.  Use at your own risk.

+++ backup.sh
@@ -14,18 +14,36 @@ BACKUPDIR='/media/backups'
 # ARCHIVEDIR='/media/archives'
 ARCHIVEDIR='/media/archives'
 
-echo "Starting backup..."
-sync
+function osbackup() {
+    echo "Performing a dry run.  If this looks OK, you'll need to confirm it."
+    sleep 3
 
-echo "rsyncing / into ${BACKUPDIR} ..."
-rsync -aAX --delete --exclude={"/dev/*","/proc/*","/sys/*","/tmp/*","/run/*","/mnt/*","/media/*","/lost+found"} / ${BACKUPDIR}
-sync
+    rsync -aAX --dry-run --delete --exclude={"/dev/*","/proc/*","/sys/*","/tmp/*","/run/*","/mnt/*","/media/*","/lost+found"} / ${BACKUPDIR}
+    read -p "Did the above rsync look OK? [y/n]: " response
 
-echo "rsync complete and synced. Tarring..."
-tar zcf ${ARCHIVEDIR}/system-backup-$(date -u -I'minutes' | cut -d+ -f1).tar.gz ${BACKUPDIR}
-sync
+    if [[ "${response,,}" != "y" ]]
+    then
+        echo "Not syncing."
+        return 1
+    else
+        echo "Starting backup..."
+        sync
+        echo "rsyncing / into ${BACKUPDIR} ..."
+        rsync -aAX --delete --exclude={"/dev/*","/proc/*","/sys/*","/tmp/*","/run/*","/mnt/*","/media/*","/lost+found"} / ${BACKUPDIR}
+        sync
+        echo "rsync complete and writes flushed to disk."
+    fi
 
-echo "tar complete and synced."
-ls -alhSr ${ARCHIVEDIR}/system-backup-*.tar.gz
-
-echo "Backup completed."
+    echo "Do you want to create a backup archive [y/n]: " response
+    if [[ "${response,,}" != "y" ]]
+    then
+        echo "Not backing up."
+        return 1
+    else
+        tar zcf ${ARCHIVEDIR}/system-backup-$(date -u -I'minutes' | cut -d+ -f1).tar.gz ${BACKUPDIR}
+        sync
+        echo "tar complete and writes flushed to disk."
+        ls -alhSr ${ARCHIVEDIR}/system-backup-*.tar.gz
+        echo "Backup completed."
+    fi
+}
